Avoid 'duplicate' orders and subscriptions, and duplicate invoices on Xero ...
Due to how SubscriptionsController works, aborting a subscription on the payment screen and recommencing would create a new subscription and order every time.
This could lead to double, triple and more records of essentially the same subscription. This really became a problem when we started using Xero, as this created duplicate invoices too..

Readers now have a method 'threatening_duplicate_subscription' to detect and prevent this scenario.
On SubscriptionsController#create, we use it to update the subscription, order and Xero invoice, rather than creating new ones.
Orders therefore have also gotten a new database column 'needs_xero_update', so that we can mark 'dirty' orders if Xero is unreachable at the time of updating records.