Stop trying to guess subscription ends_on ...
On the frontend, for non-admin visitors, we were guessing the subscription term (end date) but also not showing it.
When upgrading a subscription, this would always set the end date to end of next year, often erroneously..

After this patch, term is set correctly for new subscriptions by after_initialize, and left alone when upgrading a subscription.